Fixed weird progress arrow rendering probably

+++ src/main/java/com/blakebr0/extendedcrafting/client/gui/GuiCompressor.java
@@ -53,7 +53,7 @@ public class GuiCompressor extends GuiContainer {
 	private int getProgressBarScaled(int pixels) {
 		int i = this.tile.getProgress();
 		int j = this.tile.getRecipe().getPowerCost();
-		return j != 0 && i != 0 ? i * pixels / j : 0;
+		return (int) (j != 0 && i != 0 ? (long) i * pixels / j : 0);
 	}
 
 	private void drawItemStack(ItemStack stack, int x, int y, String altText) {
