+++ src/main/java/com/blakebr0/extendedcrafting/client/gui/GuiCraftingCore.java
@@ -38,7 +38,7 @@ public class GuiCraftingCore extends GuiContainer {
 	private int getProgressBarScaled(int pixels) {
 		int i = this.tile.getProgress();
 		int j = this.tile.getRecipe().getCost();
-		return j != 0 && i != 0 ? i * pixels / j : 0;
+		return (int) (j != 0 && i != 0 ? (long) i * pixels / j : 0);
 	}
 
 	private void drawItemStack(ItemStack stack, int x, int y, String altText) {
